+++ src/react-components/src/index.ts
@@ -1,5 +1,8 @@
 export { default as Button } from "./components/01-atoms/Button/Button";
 export { default as BodyText } from "./components/01-atoms/Text/Text/BodyText";
-export { default as Breadcrumb } from "./components/02-molecules/menus/Breadcrumbs/Breadcrumbs";
 export { default as UnderlineLink } from "./components/01-atoms/Links/UnderlineLink";
 export { default as IconLink } from "./components/01-atoms/Links/IconLink";
+export { default as Breadcrumbs } from "./components/02-molecules/Menus/Breadcrumbs/Breadcrumbs";
+export { default as IconLinkList } from "./components/03-organisms/IconLinkList/IconLinkList";
+export { default as PageHeader } from "./components/03-organisms/Headers/HeaderImgRight";
+export { default as SearchPromo } from "./components/03-organisms/SearchPromo/SearchPromo";