+++ src/react-components/CHANGELOG.md
@@ -11,9 +11,12 @@ Currently, this repo is in Prerelease.  When it is released, this project will a
 - `Image` for 2:1 ratio
 - `PageTitle`
 - `Header with Image Right`
+- `FormDropdown`
+- `SearchPromo`
 
 ### Changed 
 - Moved heading-related atoms from `01-atoms/Text` into `01-atoms/Text/Heading` 
+- Updated `Searchbar` component to include Parameters
 
 ## [0.0.5] - 2019-12-06
 ### Changed